% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{[}
\alias{[}
\alias{[,SpatialNetwork-method}
\alias{[<-}
\alias{[<-,SpatialNetwork-method}
\title{Extract or replace parts of a SpatialNetwork object}
\description{
Extract or replace parts of a SpatialNetwork object

set parts of SpatialNetwork
}

