% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{graph.barplot.bound.lower}
\alias{graph.barplot.bound.lower}
\alias{graph.barplot.bound.lower,SpatialNetwork-method}
\alias{graph.barplot.bound.lower<-,SpatialNetwork,numeric-method}
\title{Get the barplot lower bound position of a \code{SpatialNetwork} object}
\usage{
graph.barplot.bound.lower(object)

\S4method{graph.barplot.bound.lower}{SpatialNetwork}(object)

\S4method{graph.barplot.bound.lower}{SpatialNetwork,numeric}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a numeric vector of coordinates, c(x,y), specifying a shift from the center of each country.}
}
\description{
This generic method intends to extract the barplot lower bound position of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = numeric}: method for \code{SpatialNetwork} objects.
}}

