% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{graph.barplot.bound.upper<-}
\alias{graph.barplot.bound.upper<-}
\title{Set the barplot upper bound position  of a \code{SpatialNetwork} object}
\usage{
graph.barplot.bound.upper(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a numeric vector of coordinates, c(x,y), specifying a shift from the center of each country.}
}
\description{
This generic method intends to set or replace the barplot upper bound position of a \code{SpatialNetwork} object.
}

