% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{graph.barplot.fgcolor<-}
\alias{graph.barplot.fgcolor<-}
\title{Set the barplot foreground color  of a \code{SpatialNetwork} object}
\usage{
graph.barplot.fgcolor(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the color.}
}
\description{
This generic method intends to set or replace the barplot foreground color of a \code{SpatialNetwork} object.
}

