% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{graph.barplot.variable<-}
\alias{graph.barplot.variable<-}
\title{Set the barplot variable  of a \code{SpatialNetwork} object}
\usage{
graph.barplot.variable(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the name of the variable to use for plotting barplots.}
}
\description{
This generic method intends to set or replace the barplot variable of a \code{SpatialNetwork} object.
}

