% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{graph.barplot.width<-}
\alias{graph.barplot.width<-}
\title{Set the barplot width  of a \code{SpatialNetwork} object}
\usage{
graph.barplot.width(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a numeric.}
}
\description{
This generic method intends to set or replace the barplot width of a \code{SpatialNetwork} object.
}

