% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{graph.barplot.width}
\alias{graph.barplot.width}
\alias{graph.barplot.width,SpatialNetwork-method}
\alias{graph.barplot.width<-,SpatialNetwork,numeric-method}
\title{Get the barplot width of a \code{SpatialNetwork} object}
\usage{
graph.barplot.width(object)

\S4method{graph.barplot.width}{SpatialNetwork}(object)

\S4method{graph.barplot.width}{SpatialNetwork,numeric}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a numeric.}
}
\description{
This generic method intends to extract the barplot width of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = numeric}: method for \code{SpatialNetwork} objects.
}}

