% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{graph.blackwhite.enable<-}
\alias{graph.blackwhite.enable<-}
\title{Set the black and white mode status  of a \code{SpatialNetwork} object}
\usage{
graph.blackwhite.enable(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a \code{logical}, the black and white mode status.}
}
\description{
This generic method intends to set or replace the black and white mode status of a \code{SpatialNetwork} object.
}

