% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{graph.blackwhite.enable}
\alias{graph.blackwhite.enable}
\alias{graph.blackwhite.enable,SpatialNetwork-method}
\alias{graph.blackwhite.enable<-,SpatialNetwork,logical-method}
\title{Get the black and white mode status of a \code{SpatialNetwork} object}
\usage{
graph.blackwhite.enable(object)

\S4method{graph.blackwhite.enable}{SpatialNetwork}(object)

\S4method{graph.blackwhite.enable}{SpatialNetwork,logical}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a \code{logical}, the black and white mode status.}
}
\description{
This generic method intends to extract the black and white mode status of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = logical}: method for \code{SpatialNetwork} objects.
}}

