% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{graph.blackwhite.list<-}
\alias{graph.blackwhite.list<-}
\title{Set the list of all black and white mode parameters of a \code{SpatialNetwork} object}
\usage{
graph.blackwhite.list(object) <- value
}
\arguments{
\item{object}{the \code{SpatialNetwork} object for which we want to set parameters.}

\item{value}{a list of parameters.}
}
\description{
This generic method intends to set or replace black and white mode parameters of a \code{SpatialNetwork} object.
}

