% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{graph.blackwhite.list}
\alias{graph.blackwhite.list}
\alias{graph.blackwhite.list,SpatialNetwork-method}
\alias{graph.blackwhite.list<-,SpatialNetwork,list-method}
\title{Get the list of all black and white mode parameters of a \code{SpatialNetwork} object}
\usage{
graph.blackwhite.list(object)

\S4method{graph.blackwhite.list}{SpatialNetwork}(object)

\S4method{graph.blackwhite.list}{SpatialNetwork,list}(object) <- value
}
\arguments{
\item{object}{the \code{SpatialNetwork} object for which we want to get parameters.}

\item{value}{a list of parameters.}
}
\description{
This generic method intends to extract black and white mode parameters of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = list}: method for \code{SpatialNetwork} objects.
}}

