% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{graph.blackwhite.min<-}
\alias{graph.blackwhite.min<-}
\title{Set the black and white mode minimal gray value of a \code{SpatialNetwork} object}
\usage{
graph.blackwhite.min(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a \code{numeric}, the black and white mode minimal gray value.}
}
\description{
This generic method intends to set or replace the black and white mode minimal gray value (from 0 to 1) of a \code{SpatialNetwork} object.
}

