% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{graph.blackwhite.min}
\alias{graph.blackwhite.min}
\alias{graph.blackwhite.min,SpatialNetwork-method}
\alias{graph.blackwhite.min<-,SpatialNetwork,numeric-method}
\title{Get the black and white mode minimal gray value of a \code{SpatialNetwork} object}
\usage{
graph.blackwhite.min(object)

\S4method{graph.blackwhite.min}{SpatialNetwork}(object)

\S4method{graph.blackwhite.min}{SpatialNetwork,numeric}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a \code{logical}, the black and white mode minimal gray value.}
}
\description{
This generic method intends to extract the black and white mode minimal gray value (from 0 to 1) of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = numeric}: method for \code{SpatialNetwork} objects.
}}

