% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{graph.color.region}
\alias{graph.color.region}
\alias{graph.color.region,SpatialNetwork-method}
\alias{graph.color.region<-,SpatialNetwork,character-method}
\title{Get the default color of a region of a \code{SpatialNetwork} object}
\usage{
graph.color.region(object)

\S4method{graph.color.region}{SpatialNetwork}(object)

\S4method{graph.color.region}{SpatialNetwork,character}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a \code{character}, the color.}
}
\description{
This generic method intends to extract the default color of a region of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = character}: method for \code{SpatialNetwork} objects.
}}

