% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{graph.color.variable<-}
\alias{graph.color.variable<-}
\title{Set the color variable  of a \code{SpatialNetwork} object}
\usage{
graph.color.variable(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the new color, for example "#000000".}
}
\description{
This generic method intends to set or replace the color variable of a \code{SpatialNetwork} object.
}

