% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{graph.label.color<-}
\alias{graph.label.color<-}
\title{Set the label color  of a \code{SpatialNetwork} object}
\usage{
graph.label.color(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the new label, for example "#000000".}
}
\description{
This generic method intends to set or replace the label color of a \code{SpatialNetwork} object.
}

