% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{graph.legend.cex<-}
\alias{graph.legend.cex<-}
\title{Set the legend cex parameter of a \code{SpatialNetwork} object}
\usage{
graph.legend.cex(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a numeric.}
}
\description{
This generic method intends to set or replace the legend cex parameter of a \code{SpatialNetwork} object.
}

