% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{graph.legend.ncol}
\alias{graph.legend.ncol}
\alias{graph.legend.ncol,SpatialNetwork-method}
\alias{graph.legend.ncol<-,SpatialNetwork,numeric-method}
\title{Get the legend number of columns of a \code{SpatialNetwork} object}
\usage{
graph.legend.ncol(object)

\S4method{graph.legend.ncol}{SpatialNetwork}(object)

\S4method{graph.legend.ncol}{SpatialNetwork,numeric}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a numeric.}
}
\description{
This generic method intends to extract the legend number of columns of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = numeric}: method for \code{SpatialNetwork} objects.
}}

