% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{graph.legend.print<-}
\alias{graph.legend.print<-}
\title{Set the legend print (yes/no) status of a \code{SpatialNetwork} object}
\usage{
graph.legend.print(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a logical.}
}
\description{
This generic method intends to set or replace the legend print (yes/no) status of a \code{SpatialNetwork} object.
}

