% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{graph.map}
\alias{graph.map}
\alias{graph.map,SpatialNetwork-method}
\alias{graph.map<-,SpatialNetwork,SpatialPolygons-method}
\title{Get the map to a \code{SpatialNetwork} object}
\usage{
graph.map(object)

\S4method{graph.map}{SpatialNetwork}(object)

\S4method{graph.map}{SpatialNetwork,SpatialPolygons}(object) <- value
}
\arguments{
\item{object}{the \code{SpatialNetwork} object for which we want to get the map.}

\item{value}{the map.}
}
\description{
This generic method intends to extract the map object. Currently only \code{SpatialPolygons} from the \code{sp} package are supported.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialPolygons} objects.

\item \code{object = SpatialNetwork,value = SpatialPolygons}: method for \code{SpatialPolygons} objects.
}}

