% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{graph.network.arrow.head.lth}
\alias{graph.network.arrow.head.lth}
\alias{graph.network.arrow.head.lth,SpatialNetwork,character-method}
\alias{graph.network.arrow.head.lth<-,SpatialNetwork,character,numeric-method}
\title{Get the arrow head length of a given network of a \code{SpatialNetwork} object}
\usage{
graph.network.arrow.head.lth(object, network.name)

\S4method{graph.network.arrow.head.lth}{SpatialNetwork,character}(object,
  network.name)


  \S4method{graph.network.arrow.head.lth}{SpatialNetwork,character,numeric}(object,
  network.name) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{network.name}{character; the name of the network.}

\item{value}{the arrow head length.}
}
\description{
This generic method intends to extract the arrow head length of a given network of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = SpatialNetwork,network.name = character}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,network.name = character,value = numeric}: method for \code{SpatialNetwork} objects.
}}

