% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{graph.network.arrow.head.type<-}
\alias{graph.network.arrow.head.type<-}
\title{Set the arrow head type of a given network of a \code{SpatialNetwork} object}
\usage{
graph.network.arrow.head.type(object, network.name) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{network.name}{character; the name of the network.}

\item{value}{type of arrowhead to draw, one of "simple", "curved", "triangle", "circle", "ellipse" or "T". See \code{\link[shape]{Arrows}} for details.}
}
\description{
This generic method intends to set or replace the arrow head type of a given network of a \code{SpatialNetwork} object.
}

