% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{graph.network.arrow.line.type<-}
\alias{graph.network.arrow.line.type<-}
\title{Set the arrow line type of a given network of a \code{SpatialNetwork} object}
\usage{
graph.network.arrow.line.type(object, network.name) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{network.name}{character; the name of the network.}

\item{value}{a numeric; the arrow line type.}
}
\description{
This generic method intends to set or replace the arrow line type of a given network of a \code{SpatialNetwork} object.
}

