% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{graph.network.exists}
\alias{graph.network.exists}
\title{Test if a network exist}
\usage{
graph.network.exists(object, network.name)
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{network.name}{a character; the name of the network.}
}
\description{
This function tests if the network name given in parameter match the name of a network defined within a \code{SpatialNetwork} object.
}

