% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{graph.network.label<-}
\alias{graph.network.label<-}
\title{Set the label of a given network of a \code{SpatialNetwork} object}
\usage{
graph.network.label(object, network.name) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{network.name}{character; the name of the network.}

\item{value}{the network label.}
}
\description{
This generic method intends to set or replace the label of a given network of a \code{SpatialNetwork} object.
}

