% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{graph.network.label}
\alias{graph.network.label}
\alias{graph.network.label,SpatialNetwork,character-method}
\alias{graph.network.label<-,SpatialNetwork,character,character-method}
\title{Get the label of a given network of a \code{SpatialNetwork} object}
\usage{
graph.network.label(object, network.name)

\S4method{graph.network.label}{SpatialNetwork,character}(object, network.name)

\S4method{graph.network.label}{SpatialNetwork,character,character}(object,
  network.name) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{network.name}{character; the name of the network.}

\item{value}{the network label.}
}
\description{
This generic method intends to extract the label of a given network of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = SpatialNetwork,network.name = character}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,network.name = character,value = character}: method for \code{SpatialNetwork} objects.
}}

