% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{graph.networks.add<-}
\alias{graph.networks.add<-}
\alias{graph.networks.add<-,SpatialNetwork,character-method}
\title{Add a network}
\usage{
graph.networks.add(object) <- value

\S4method{graph.networks.add}{SpatialNetwork,character}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a character; the name of the network.}
}
\description{
This function defines a new network item in a \code{SpatialNetwork} object.
}

