% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{graph.networks.remove<-}
\alias{graph.networks.remove<-}
\alias{graph.networks.remove<-,SpatialNetwork,character-method}
\title{Remove a network}
\usage{
graph.networks.remove(object) <- value

\S4method{graph.networks.remove}{SpatialNetwork,character}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a character; the name of the network.}
}
\description{
This function remove a network item in a \code{SpatialNetwork} object.
}

