% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{graph.symbol.cex<-}
\alias{graph.symbol.cex<-}
\title{Set the symbol cex parameter  of a \code{SpatialNetwork} object}
\usage{
graph.symbol.cex(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the new cex parameter.}
}
\description{
This generic method intends to set or replace the symbol cex parameter of a \code{SpatialNetwork} object.
}

