% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{graph.symbol.cex}
\alias{graph.symbol.cex}
\alias{graph.symbol.cex,SpatialNetwork-method}
\alias{graph.symbol.cex<-,SpatialNetwork,numeric-method}
\title{Get the symbol cex parameter of a \code{SpatialNetwork} object}
\usage{
graph.symbol.cex(object)

\S4method{graph.symbol.cex}{SpatialNetwork}(object)

\S4method{graph.symbol.cex}{SpatialNetwork,numeric}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the new cex parameter.}
}
\description{
This generic method intends to extract the symbol cex parameter of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = numeric}: method for \code{SpatialNetwork} objects.
}}

