% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{graph.symbol.color<-}
\alias{graph.symbol.color<-}
\title{Set the symbol color of a \code{SpatialNetwork} object}
\usage{
graph.symbol.color(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the color.}
}
\description{
This generic method intends to set or replace the symbol color of a \code{SpatialNetwork} object.
}

