% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{graph.symbol.legend<-}
\alias{graph.symbol.legend<-}
\title{Set the symbol legend  of a \code{SpatialNetwork} object}
\usage{
graph.symbol.legend(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the new legend.}
}
\description{
This generic method intends to set or replace the symbol legend of a \code{SpatialNetwork} object.
}

