% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{graph.symbol.variable<-}
\alias{graph.symbol.variable<-}
\title{Set the symbol variable  of a \code{SpatialNetwork} object}
\usage{
graph.symbol.variable(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the symbol variable.}
}
\description{
This generic method intends to set or replace the symbol variable of a \code{SpatialNetwork} object.
}

