% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{graph.symbol.variable}
\alias{graph.symbol.variable}
\alias{graph.symbol.variable,SpatialNetwork-method}
\alias{graph.symbol.variable<-,SpatialNetwork,character-method}
\title{Get the symbol variable of a \code{SpatialNetwork} object}
\usage{
graph.symbol.variable(object)

\S4method{graph.symbol.variable}{SpatialNetwork}(object)

\S4method{graph.symbol.variable}{SpatialNetwork,character}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the symbol variable.}
}
\description{
This generic method intends to extract the symbol variable of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = character}: method for \code{SpatialNetwork} objects.
}}

