% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{graph.title.sub<-}
\alias{graph.title.sub<-}
\title{Set the sub title  of a \code{SpatialNetwork} object}
\usage{
graph.title.sub(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the new title.}
}
\description{
This generic method intends to set or replace the sub title of a \code{SpatialNetwork} object.
}

