% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{graph.title.sub}
\alias{graph.title.sub}
\alias{graph.title.sub,SpatialNetwork-method}
\alias{graph.title.sub<-,SpatialNetwork,character-method}
\title{Get the sub title of a \code{SpatialNetwork} object}
\usage{
graph.title.sub(object)

\S4method{graph.title.sub}{SpatialNetwork}(object)

\S4method{graph.title.sub}{SpatialNetwork,character}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the new title.}
}
\description{
This generic method intends to extract the sub title of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = character}: method for \code{SpatialNetwork} objects.
}}

