\name{plot.position-methods}
\docType{methods}
\alias{plot.position-methods}
\alias{plot.position,SpatialNetwork-method}
\alias{plot.position,SpatialPolygons-method}
\title{Methods for Function \code{plot.position} in Package \pkg{spnet}}
\description{
Methods for function \code{plot.position} in package \pkg{spnet}
}
\section{Methods}{
\describe{

\item{\code{signature(x = "SpatialNetwork")}}{
\code{plot.position} method for \code{SpatialNetwork} objects.
}

\item{\code{signature(x = "SpatialPolygons")}}{
\code{plot.position} method for \code{SpatialPolygons} objects.
}
}}
\seealso{
\code{\link{plot.position}} function.
}
\keyword{methods}
\keyword{ plot }
\keyword{ spatial }
\keyword{ SpatialPolygons }
\keyword{ SpatialNetwork }