\name{spnet-package}
\alias{spnet-package}
\alias{spnet}
\docType{package}
\title{
Plotting social networks on maps
}
\description{
The spnet package offers methods for dealing with spacial social networks. It allows to plot networks for which actors have a specific location on a map (participants in a political debate, cities, etc.). SpatialPolygons objects from the sp package are supported.
}
\details{
\tabular{ll}{
Package: \tab spnet\cr
Type: \tab Package\cr
Version: \tab 0.1-1\cr
Date: \tab 2013-07-12\cr
License: \tab GPL-3\cr
}
~~ An overview of how to use the package, ~~
~~ including the most important functions ~~
}
\author{
Emmanuel Rousseaux, Marion Deville-Naggay

Maintainer: Emmanuel Rousseaux <emmanuel.rousseaux@unige.ch>
}
\references{
Deville-Naggay M. and Rousseaux E. (2013). TITLE. JOURNAL. VOLUME(NUMBER), PAGEFROM-PAGETO.
}
\keyword{ package }
\keyword{ spnet }
\keyword{ networks }
\keyword{ map }
\keyword{ spatial }

\examples{

}
