% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spnet.barplot.bgcolor<-}
\alias{spnet.barplot.bgcolor<-}
\title{Set the barplot background color  of a \code{SpatialNetwork} object}
\usage{
spnet.barplot.bgcolor(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the new color.}
}
\description{
This generic method intends to set or replace the barplot background color of a \code{SpatialNetwork} object.
}

