% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{spnet.barplot.bgcolor}
\alias{spnet.barplot.bgcolor}
\alias{spnet.barplot.bgcolor,SpatialNetwork-method}
\alias{spnet.barplot.bgcolor<-,SpatialNetwork,character-method}
\title{Get the barplot background color of a \code{SpatialNetwork} object}
\usage{
spnet.barplot.bgcolor(object)

\S4method{spnet.barplot.bgcolor}{SpatialNetwork}(object)

\S4method{spnet.barplot.bgcolor}{SpatialNetwork,character}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the new color.}
}
\description{
This generic method intends to extract the barplot background color of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = character}: method for \code{SpatialNetwork} objects.
}}

