% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spnet.barplot.bound.lower<-}
\alias{spnet.barplot.bound.lower<-}
\title{Set the barplot lower bound position  of a \code{SpatialNetwork} object}
\usage{
spnet.barplot.bound.lower(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a numeric vector of coordinates, c(x,y), specifying a translation from the center of each country.}
}
\description{
This generic method intends to add or replace the barplot lower bound position of a \code{SpatialNetwork} object.
}

