% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{spnet.barplot.fgcolor}
\alias{spnet.barplot.fgcolor}
\alias{spnet.barplot.fgcolor,SpatialNetwork-method}
\alias{spnet.barplot.fgcolor<-,SpatialNetwork,character-method}
\title{Get the barplot foreground color of a \code{SpatialNetwork} object}
\usage{
spnet.barplot.fgcolor(object)

\S4method{spnet.barplot.fgcolor}{SpatialNetwork}(object)

\S4method{spnet.barplot.fgcolor}{SpatialNetwork,character}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the color.}
}
\description{
This generic method intends to extract the barplot foreground color of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = character}: method for \code{SpatialNetwork} objects.
}}

