% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{spnet.barplot.list}
\alias{spnet.barplot.list}
\alias{spnet.barplot.list,SpatialNetwork-method}
\alias{spnet.barplot.list<-,SpatialNetwork,list-method}
\title{Get the list of all barplot parameters of a \code{SpatialNetwork} object}
\usage{
spnet.barplot.list(object)

\S4method{spnet.barplot.list}{SpatialNetwork}(object)

\S4method{spnet.barplot.list}{SpatialNetwork,list}(object) <- value
}
\arguments{
\item{object}{the \code{SpatialNetwork} object for which we want to get parameters.}

\item{value}{a list of parameters.}
}
\description{
This generic method intends to extract barplot parameters of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = list}: method for \code{SpatialNetwork} objects.
}}

