% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{spnet.barplot.variable}
\alias{spnet.barplot.variable}
\alias{spnet.barplot.variable,SpatialNetwork-method}
\alias{spnet.barplot.variable<-,SpatialNetwork,character-method}
\title{Get the barplot variable of a \code{SpatialNetwork} object}
\usage{
spnet.barplot.variable(object)

\S4method{spnet.barplot.variable}{SpatialNetwork}(object)

\S4method{spnet.barplot.variable}{SpatialNetwork,character}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the name of the variable to use for plotting barplots.}
}
\description{
This generic method intends to extract the barplot variable of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = character}: method for \code{SpatialNetwork} objects.
}}

