% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{spnet.color.background<-}
\alias{spnet.color.background<-}
\title{Set the background color  of a \code{SpatialNetwork} object}
\usage{
spnet.color.background(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a \code{character}, the color.}
}
\description{
This generic method intends to set or replace the background color of a \code{SpatialNetwork} object.
}

