% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{spnet.color.background}
\alias{spnet.color.background}
\alias{spnet.color.background,SpatialNetwork-method}
\alias{spnet.color.background<-,SpatialNetwork,character-method}
\title{Get the background color of a \code{SpatialNetwork} object}
\usage{
spnet.color.background(object)

\S4method{spnet.color.background}{SpatialNetwork}(object)

\S4method{spnet.color.background}{SpatialNetwork,character}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a \code{character}, the color.}
}
\description{
This generic method intends to extract the background color of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = character}: method for \code{SpatialNetwork} objects.
}}

