% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{spnet.color.border<-}
\alias{spnet.color.border<-}
\title{Set the border color  of a \code{SpatialNetwork} object}
\usage{
spnet.color.border(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a \code{character}, the color.}
}
\description{
This generic method intends to set or replace the border color of a \code{SpatialNetwork} object.
}

