% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{spnet.color.border}
\alias{spnet.color.border}
\alias{spnet.color.border,SpatialNetwork-method}
\alias{spnet.color.border<-,SpatialNetwork,character-method}
\title{Get the border color of a \code{SpatialNetwork} object}
\usage{
spnet.color.border(object)

\S4method{spnet.color.border}{SpatialNetwork}(object)

\S4method{spnet.color.border}{SpatialNetwork,character}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a \code{character}, the color.}
}
\description{
This generic method intends to extract the border color of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = character}: method for \code{SpatialNetwork} objects.
}}

