% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{spnet.color.legend<-}
\alias{spnet.color.legend<-}
\title{Set the color legend  of a \code{SpatialNetwork} object}
\usage{
spnet.color.legend(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the color legend.}
}
\description{
This generic method intends to set or replace the color legend of a \code{SpatialNetwork} object.
}

