% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spnet.color.variable<-}
\alias{spnet.color.variable<-}
\title{Set the color variable  of a \code{SpatialNetwork} object}
\usage{
spnet.color.variable(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the new color, for example "#000000".}
}
\description{
This generic method intends to set or replace the color variable of a \code{SpatialNetwork} object.
}

