% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{spnet.color.variable}
\alias{spnet.color.variable}
\alias{spnet.color.variable,SpatialNetwork-method}
\alias{spnet.color.variable<-,SpatialNetwork,character-method}
\title{Get the color variable of a \code{SpatialNetwork} object}
\usage{
spnet.color.variable(object)

\S4method{spnet.color.variable}{SpatialNetwork}(object)

\S4method{spnet.color.variable}{SpatialNetwork,character}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the new color, for example "#000000".}
}
\description{
This generic method intends to extract the color variable of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = character}: method for \code{SpatialNetwork} objects.
}}

