% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spnet.example.basic.r
\name{spnet.example.basic}
\alias{spnet.example.basic}
\alias{spnet.example.basic.full}
\alias{spnet.example.basic.map}
\title{Spnet basic examples}
\usage{
spnet.example.basic(map = TRUE, color = TRUE, symbol = TRUE,
  network1 = TRUE, network2 = TRUE, barplot = TRUE, title = TRUE)

spnet.example.basic.full()

spnet.example.basic.map()
}
\arguments{
\item{map}{logical; if \code{TRUE} an example of map is provided.}

\item{color}{logical; if \code{TRUE} an example of map colorization is provided.}

\item{symbol}{logical; if \code{TRUE} an example of symbol use is provided.}

\item{network1}{logical; if \code{TRUE} a first example of network is provided.}

\item{network2}{logical; if \code{TRUE} a second example of network is provided.}

\item{barplot}{logical; if \code{TRUE} a example of barplot rendering of a numeric variable is provided.}

\item{title}{logical; if \code{TRUE} a example of title is provided.}
}
\value{
a \code{SpatialNetwok} object.
}
\description{
Create SpatialNetwork object examples for demonstration and testing purpose.
}
\examples{
data(world.map.simplified, package = "spnet")
net1 <- spnet.example.basic()
plot(net1)
}

