% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tools.global.r
\name{spnet.get.local.user.manual}
\alias{spnet.get.local.user.manual}
\title{get the local copy of the spnet user manual}
\usage{
spnet.get.local.user.manual(where = getwd(), overwrite = FALSE)
}
\arguments{
\item{where}{the location where to copy the user manual. Default is the working directory.}

\item{overwrite}{logical; should existing destination files be overwritten?}
}
\description{
This function copies the spnet user manual to a user defined directory.
}

