% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spnet.label.color<-}
\alias{spnet.label.color<-}
\title{Set the label color  of a \code{SpatialNetwork} object}
\usage{
spnet.label.color(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the new label, for example "#000000".}
}
\description{
This generic method intends to add or replace the label color of a \code{SpatialNetwork} object.
}

