% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{spnet.label.color}
\alias{spnet.label.color}
\alias{spnet.label.color,SpatialNetwork-method}
\alias{spnet.label.color<-,SpatialNetwork,character-method}
\title{Get the label color of a \code{SpatialNetwork} object}
\usage{
spnet.label.color(object)

\S4method{spnet.label.color}{SpatialNetwork}(object)

\S4method{spnet.label.color}{SpatialNetwork,character}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the new label, for example "#000000".}
}
\description{
This generic method intends to extract the label color of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = character}: method for \code{SpatialNetwork} objects.
}}

