% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{spnet.label.list<-}
\alias{spnet.label.list<-}
\title{Set the list of all label parameters of a \code{SpatialNetwork} object}
\usage{
spnet.label.list(object) <- value
}
\arguments{
\item{object}{the \code{SpatialNetwork} object for which we want to set parameters.}

\item{value}{a list of parameters.}
}
\description{
This generic method intends to set or replace label parameters of a \code{SpatialNetwork} object.
}

