% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{spnet.label.variable<-}
\alias{spnet.label.variable<-}
\title{Set the label variable  of a \code{SpatialNetwork} object}
\usage{
spnet.label.variable(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the new label, for example "#000000".}
}
\description{
This generic method intends to set or replace the label variable of a \code{SpatialNetwork} object.
}

