% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{spnet.label.variable}
\alias{spnet.label.variable}
\alias{spnet.label.variable,SpatialNetwork-method}
\alias{spnet.label.variable<-,SpatialNetwork,character-method}
\title{Get the label variable of a \code{SpatialNetwork} object}
\usage{
spnet.label.variable(object)

\S4method{spnet.label.variable}{SpatialNetwork}(object)

\S4method{spnet.label.variable}{SpatialNetwork,character}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the new label, for example "#000000".}
}
\description{
This generic method intends to extract the label variable of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = character}: method for \code{SpatialNetwork} objects.
}}

