% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spnet.legend.cex<-}
\alias{spnet.legend.cex<-}
\title{Set the legend cex parameter of a \code{SpatialNetwork} object}
\usage{
spnet.legend.cex(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a numeric.}
}
\description{
This generic method intends to add or replace the legend cex parameter of a \code{SpatialNetwork} object.
}

