% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{spnet.legend.horiz<-}
\alias{spnet.legend.horiz<-}
\title{Set the legend horizontal or vertical setting of a \code{SpatialNetwork} object}
\usage{
spnet.legend.horiz(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a logical.}
}
\description{
This generic method intends to set or replace the legend horizontal or vertical setting of a \code{SpatialNetwork} object.
}

