% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{spnet.legend.horiz}
\alias{spnet.legend.horiz}
\alias{spnet.legend.horiz,SpatialNetwork-method}
\alias{spnet.legend.horiz<-,SpatialNetwork,logical-method}
\title{Get the legend horizontal or vertical setting of a \code{SpatialNetwork} object}
\usage{
spnet.legend.horiz(object)

\S4method{spnet.legend.horiz}{SpatialNetwork}(object)

\S4method{spnet.legend.horiz}{SpatialNetwork,logical}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a logical.}
}
\description{
This generic method intends to extract the legend horizontal or vertical setting of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = logical}: method for \code{SpatialNetwork} objects.
}}

