% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{spnet.legend.line.width}
\alias{spnet.legend.line.width}
\alias{spnet.legend.line.width,SpatialNetwork-method}
\alias{spnet.legend.line.width<-,SpatialNetwork,numeric-method}
\title{Get the legend line width parameter of a \code{SpatialNetwork} object}
\usage{
spnet.legend.line.width(object)

\S4method{spnet.legend.line.width}{SpatialNetwork}(object)

\S4method{spnet.legend.line.width}{SpatialNetwork,numeric}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a logical.}
}
\description{
This generic method intends to extract the legend line width parameter of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = numeric}: method for \code{SpatialNetwork} objects.
}}

