% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spnet.legend.print<-}
\alias{spnet.legend.print<-}
\title{Set the legend print (yes/no) status of a \code{SpatialNetwork} object}
\usage{
spnet.legend.print(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a logical.}
}
\description{
This generic method intends to add or replace the legend print (yes/no) status of a \code{SpatialNetwork} object.
}

