% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{spnet.legend.print}
\alias{spnet.legend.print}
\alias{spnet.legend.print,SpatialNetwork-method}
\alias{spnet.legend.print<-,SpatialNetwork,logical-method}
\title{Get the legend print (yes/no) status of a \code{SpatialNetwork} object}
\usage{
spnet.legend.print(object)

\S4method{spnet.legend.print}{SpatialNetwork}(object)

\S4method{spnet.legend.print}{SpatialNetwork,logical}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a logical.}
}
\description{
This generic method intends to extract the legend print (yes/no) status of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = logical}: method for \code{SpatialNetwork} objects.
}}

