% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spnet.map<-}
\alias{spnet.map<-}
\title{Set the map to a \code{SpatialNetwork} object}
\usage{
spnet.map(object) <- value
}
\arguments{
\item{object}{the \code{SpatialNetwork} object for which we want to set the map.}

\item{value}{the map.}
}
\description{
This generic method intends to add or replace the map object. Currently only \code{SpatialPolygons} from the \code{sp} package are supported.
}

