% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{spnet.network.arrow.head.length}
\alias{spnet.network.arrow.head.length}
\alias{spnet.network.arrow.head.length,SpatialNetwork,character-method}
\alias{spnet.network.arrow.head.length<-,SpatialNetwork,character,numeric-method}
\title{Get the arrow head length of a given network of a \code{SpatialNetwork} object}
\usage{
spnet.network.arrow.head.length(object, network.name)

\S4method{spnet.network.arrow.head.length}{SpatialNetwork,character}(object,
  network.name)


  \S4method{spnet.network.arrow.head.length}{SpatialNetwork,character,numeric}(object,
  network.name) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{network.name}{character; the name of the network.}

\item{value}{the arrow head length.}
}
\description{
This generic method intends to extract the arrow head length of a given network of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = SpatialNetwork,network.name = character}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,network.name = character,value = numeric}: method for \code{SpatialNetwork} objects.
}}

