% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spnet.network.arrow.head.lth<-}
\alias{spnet.network.arrow.head.lth<-}
\title{Set the arrow head length of a given network of a \code{SpatialNetwork} object}
\usage{
spnet.network.arrow.head.lth(object, network.name) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{network.name}{character; the name of the network.}

\item{value}{the arrow head length.}
}
\description{
This generic method intends to set or replace the arrow head length of a given network of a \code{SpatialNetwork} object.
}

