% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{spnet.network.arrow.opacity}
\alias{spnet.network.arrow.opacity}
\alias{spnet.network.arrow.opacity,SpatialNetwork,character-method}
\alias{spnet.network.arrow.opacity<-,SpatialNetwork,character,numeric-method}
\title{Get the arrow opacity of a given network of a \code{SpatialNetwork} object}
\usage{
spnet.network.arrow.opacity(object, network.name)

\S4method{spnet.network.arrow.opacity}{SpatialNetwork,character}(object,
  network.name)


  \S4method{spnet.network.arrow.opacity}{SpatialNetwork,character,numeric}(object,
  network.name) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{network.name}{character; the name of the network.}

\item{value}{the arrow opacity.}
}
\description{
This generic method intends to extract the arrow opacity of a given network of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = SpatialNetwork,network.name = character}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,network.name = character,value = numeric}: method for \code{SpatialNetwork} objects.
}}

