% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{spnet.network.arrow.shift.x<-}
\alias{spnet.network.arrow.shift.x<-}
\title{Set the arrow shift on the x axis of a given network of a \code{SpatialNetwork} object}
\usage{
spnet.network.arrow.shift.x(object, network.name) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{network.name}{character; the name of the network.}

\item{value}{the arrow shift on the x axis.}
}
\description{
This generic method intends to set or replace the arrow shift on the x axis of a given network of a \code{SpatialNetwork} object.
}

