% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{spnet.network.arrow.shift.y}
\alias{spnet.network.arrow.shift.y}
\alias{spnet.network.arrow.shift.y,SpatialNetwork,character-method}
\alias{spnet.network.arrow.shift.y<-,SpatialNetwork,character,numeric-method}
\title{Get the arrow shift on the y axis of a given network of a \code{SpatialNetwork} object}
\usage{
spnet.network.arrow.shift.y(object, network.name)

\S4method{spnet.network.arrow.shift.y}{SpatialNetwork,character}(object,
  network.name)


  \S4method{spnet.network.arrow.shift.y}{SpatialNetwork,character,numeric}(object,
  network.name) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{network.name}{character; the name of the network.}

\item{value}{the arrow shift on the y axis.}
}
\description{
This generic method intends to extract the arrow shift on the y axis of a given network of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = SpatialNetwork,network.name = character}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,network.name = character,value = numeric}: method for \code{SpatialNetwork} objects.
}}

