% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{spnet.network.arrow.shortening}
\alias{spnet.network.arrow.shortening}
\alias{spnet.network.arrow.shortening,SpatialNetwork,character-method}
\alias{spnet.network.arrow.shortening<-,SpatialNetwork,character,character-method}
\title{Get the arrow shortening of a given network of a \code{SpatialNetwork} object}
\usage{
spnet.network.arrow.shortening(object, network.name)

\S4method{spnet.network.arrow.shortening}{SpatialNetwork,character}(object,
  network.name)


  \S4method{spnet.network.arrow.shortening}{SpatialNetwork,character,character}(object,
  network.name) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{network.name}{character; the name of the network.}

\item{value}{the arrow shortening.}
}
\description{
This generic method intends to extract the arrow shortening of a given network of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = SpatialNetwork,network.name = character}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,network.name = character,value = character}: method for \code{SpatialNetwork} objects.
}}

