% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{spnet.network.arrow.thickness<-}
\alias{spnet.network.arrow.thickness<-}
\title{Set the arrow thickness of a given network of a \code{SpatialNetwork} object}
\usage{
spnet.network.arrow.thickness(object, network.name) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{network.name}{character; the name of the network.}

\item{value}{the arrow thickness.}
}
\description{
This generic method intends to set or replace the arrow thickness of a given network of a \code{SpatialNetwork} object.
}

