% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{spnet.network.arrow.thickness}
\alias{spnet.network.arrow.thickness}
\alias{spnet.network.arrow.thickness,SpatialNetwork,character-method}
\alias{spnet.network.arrow.thickness<-,SpatialNetwork,character,numeric-method}
\title{Get the arrow thickness of a given network of a \code{SpatialNetwork} object}
\usage{
spnet.network.arrow.thickness(object, network.name)

\S4method{spnet.network.arrow.thickness}{SpatialNetwork,character}(object,
  network.name)


  \S4method{spnet.network.arrow.thickness}{SpatialNetwork,character,numeric}(object,
  network.name) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{network.name}{character; the name of the network.}

\item{value}{the arrow thickness.}
}
\description{
This generic method intends to extract the arrow thickness of a given network of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = SpatialNetwork,network.name = character}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,network.name = character,value = numeric}: method for \code{SpatialNetwork} objects.
}}

