% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spnet.network.arrow.translate.y<-}
\alias{spnet.network.arrow.translate.y<-}
\title{Set the arrow translation on the y axis of a given network of a \code{SpatialNetwork} object}
\usage{
spnet.network.arrow.translate.y(object, network.name) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{network.name}{character; the name of the network.}

\item{value}{the arrow translation on the y axis.}
}
\description{
This generic method intends to set or replace the arrow translation on the y axis of a given network of a \code{SpatialNetwork} object.
}

