% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{spnet.network.data<-}
\alias{spnet.network.data<-}
\title{Set the data of a given network of a \code{SpatialNetwork} object}
\usage{
spnet.network.data(object, network.name) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{network.name}{character; the name of the network.}

\item{value}{the network data. Currently only support a \code{matrix} object.}
}
\description{
This generic method intends to set or replace the data of a given network of a \code{SpatialNetwork} object.
}

