% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{spnet.network.data}
\alias{spnet.network.data}
\alias{spnet.network.data,SpatialNetwork,character-method}
\alias{spnet.network.data<-,SpatialNetwork,character,matrix-method}
\title{Get the data of a given network of a \code{SpatialNetwork} object}
\usage{
spnet.network.data(object, network.name)

\S4method{spnet.network.data}{SpatialNetwork,character}(object, network.name)

\S4method{spnet.network.data}{SpatialNetwork,character,matrix}(object,
  network.name) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{network.name}{character; the name of the network.}

\item{value}{the network data. Currently only support a \code{matrix} object.}
}
\description{
This generic method intends to extract the data of a given network of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = SpatialNetwork,network.name = character}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,network.name = character,value = matrix}: method for \code{SpatialNetwork} objects.
}}

