% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{spnet.network.exists}
\alias{spnet.network.exists}
\title{Test if a network exist}
\usage{
spnet.network.exists(object, network.name)
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{network.name}{a character; the name of the network.}
}
\description{
This function tests if the network name given in parameter match the name of a network defined within a \code{SpatialNetwork} object.
}

