% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{spnet.network.list}
\alias{spnet.network.list}
\alias{spnet.network.list,SpatialNetwork,character-method}
\alias{spnet.network.list<-,SpatialNetwork,character,list-method}
\title{Get the list of all parameters of a given network of a \code{SpatialNetwork} object}
\usage{
spnet.network.list(object, network.name)

\S4method{spnet.network.list}{SpatialNetwork,character}(object, network.name)

\S4method{spnet.network.list}{SpatialNetwork,character,list}(object,
  network.name) <- value
}
\arguments{
\item{object}{the \code{SpatialNetwork} object for which we want to get parameters.}

\item{network.name}{character; the name of the network.}

\item{value}{a list of parameters.}
}
\description{
This generic method intends to extract all parameters of a given network of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = SpatialNetwork,network.name = character}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,network.name = character,value = list}: method for \code{SpatialNetwork} objects.
}}

