% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{spnet.networks.remove<-}
\alias{spnet.networks.remove<-}
\alias{spnet.networks.remove<-,SpatialNetwork,character-method}
\title{Remove a network}
\usage{
spnet.networks.remove(object) <- value

\S4method{spnet.networks.remove}{SpatialNetwork,character}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a character; the name of the network.}
}
\description{
This function remove a network item in a \code{SpatialNetwork} object.
}

