% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spnet.par.list<-}
\alias{spnet.par.list<-}
\title{Set the list of all par parameters of a \code{SpatialNetwork} object}
\usage{
spnet.par.list(object) <- value
}
\arguments{
\item{object}{the \code{SpatialNetwork} object for which we want to set parameters.}

\item{value}{a list of parameters.}
}
\description{
This generic method intends to add or replace par parameters of a \code{SpatialNetwork} object.
}

