% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{spnet.par.list}
\alias{spnet.par.list}
\alias{spnet.par.list,SpatialNetwork-method}
\alias{spnet.par.list<-,SpatialNetwork,list-method}
\title{Get the list of all par parameters of a \code{SpatialNetwork} object}
\usage{
spnet.par.list(object)

\S4method{spnet.par.list}{SpatialNetwork}(object)

\S4method{spnet.par.list}{SpatialNetwork,list}(object) <- value
}
\arguments{
\item{object}{the \code{SpatialNetwork} object for which we want to get parameters.}

\item{value}{a list of parameters.}
}
\description{
This generic method intends to extract par parameters of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = list}: method for \code{SpatialNetwork} objects.
}}

