% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{spnet.symbol.cex<-}
\alias{spnet.symbol.cex<-}
\title{Set the symbol cex parameter  of a \code{SpatialNetwork} object}
\usage{
spnet.symbol.cex(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the new cex parameter.}
}
\description{
This generic method intends to set or replace the symbol cex parameter of a \code{SpatialNetwork} object.
}

