% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spnet.symbol.color<-}
\alias{spnet.symbol.color<-}
\title{Set the symbol color of a \code{SpatialNetwork} object}
\usage{
spnet.symbol.color(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the color.}
}
\description{
This generic method intends to set or replace the symbol color of a \code{SpatialNetwork} object.
}

