% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{spnet.symbol.color}
\alias{spnet.symbol.color}
\alias{spnet.symbol.color,SpatialNetwork-method}
\alias{spnet.symbol.color<-,SpatialNetwork,character-method}
\title{Get the color symbol of a \code{SpatialNetwork} object}
\usage{
spnet.symbol.color(object)

\S4method{spnet.symbol.color}{SpatialNetwork}(object)

\S4method{spnet.symbol.color}{SpatialNetwork,character}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the color.}
}
\description{
This generic method intends to extract the color symbol of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = character}: method for \code{SpatialNetwork} objects.
}}

