% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{spnet.symbol.legend}
\alias{spnet.symbol.legend}
\alias{spnet.symbol.legend,SpatialNetwork-method}
\alias{spnet.symbol.legend<-,SpatialNetwork,character-method}
\title{Get the symbol legend of a \code{SpatialNetwork} object}
\usage{
spnet.symbol.legend(object)

\S4method{spnet.symbol.legend}{SpatialNetwork}(object)

\S4method{spnet.symbol.legend}{SpatialNetwork,character}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the new legend.}
}
\description{
This generic method intends to extract the symbol legend of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = character}: method for \code{SpatialNetwork} objects.
}}

