% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spnet.symbol.list<-}
\alias{spnet.symbol.list<-}
\title{Set the list of all symbol parameters of a \code{SpatialNetwork} object}
\usage{
spnet.symbol.list(object) <- value
}
\arguments{
\item{object}{the \code{SpatialNetwork} object for which we want to set parameters.}

\item{value}{a list of parameters.}
}
\description{
This generic method intends to set or replace symbol parameters of a \code{SpatialNetwork} object.
}

