% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{spnet.symbol.shift.x<-}
\alias{spnet.symbol.shift.x<-}
\title{Set the symbol shift on the x axis of a \code{SpatialNetwork} object}
\usage{
spnet.symbol.shift.x(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a numeric; the value of the shift.}
}
\description{
This generic method intends to set or replace the value of symbol shift on the x axis of a \code{SpatialNetwork} object.
}

