% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{spnet.symbol.shift.x}
\alias{spnet.symbol.shift.x}
\alias{spnet.symbol.shift.x,SpatialNetwork-method}
\alias{spnet.symbol.shift.x<-,SpatialNetwork,numeric-method}
\title{Get the symbol shift on the x axis of a \code{SpatialNetwork} object}
\usage{
spnet.symbol.shift.x(object)

\S4method{spnet.symbol.shift.x}{SpatialNetwork}(object)

\S4method{spnet.symbol.shift.x}{SpatialNetwork,numeric}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a numeric; the value of the shift.s}
}
\description{
This generic method intends to extract the value of symbol shift on the x axis of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = numeric}: method for \code{SpatialNetwork} objects.
}}

