% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{spnet.symbol.shift.y<-}
\alias{spnet.symbol.shift.y<-}
\title{Set the symbol shift on the y axis of a \code{SpatialNetwork} object}
\usage{
spnet.symbol.shift.y(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a numeric; the value of the shift.}
}
\description{
This generic method intends to set or replace the value of the symbol shift on the y axis of a \code{SpatialNetwork} object.
}

