% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{spnet.symbol.shift.y}
\alias{spnet.symbol.shift.y}
\alias{spnet.symbol.shift.y,SpatialNetwork-method}
\alias{spnet.symbol.shift.y<-,SpatialNetwork,numeric-method}
\title{Get the symbol shift on the y axis of a \code{SpatialNetwork} object}
\usage{
spnet.symbol.shift.y(object)

\S4method{spnet.symbol.shift.y}{SpatialNetwork}(object)

\S4method{spnet.symbol.shift.y}{SpatialNetwork,numeric}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a numeric; the value of the shift.}
}
\description{
This generic method intends to extract the value of the symbol shift on the y of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = numeric}: method for \code{SpatialNetwork} objects.
}}

