% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spnet.symbol.translate.x<-}
\alias{spnet.symbol.translate.x<-}
\title{Set the symbol translation on the x axis of a \code{SpatialNetwork} object}
\usage{
spnet.symbol.translate.x(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a numeric; the value of the translation.}
}
\description{
This generic method intends to set or replace the value of symbol translation on the x axis of a \code{SpatialNetwork} object.
}

