% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spnet.symbol.translate.y<-}
\alias{spnet.symbol.translate.y<-}
\title{Set the symbol translation on the y axis of a \code{SpatialNetwork} object}
\usage{
spnet.symbol.translate.y(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a numeric; the value of the translation.}
}
\description{
This generic method intends to add or replace the value of the symbol translation on the y axis of a \code{SpatialNetwork} object.
}

