% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{spnet.symbol.translate.y}
\alias{spnet.symbol.translate.y}
\alias{spnet.symbol.translate.y,SpatialNetwork-method}
\alias{spnet.symbol.translate.y<-,SpatialNetwork,numeric-method}
\title{Get the symbol translation on the y axis of a \code{SpatialNetwork} object}
\usage{
spnet.symbol.translate.y(object)

\S4method{spnet.symbol.translate.y}{SpatialNetwork}(object)

\S4method{spnet.symbol.translate.y}{SpatialNetwork,numeric}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a numeric; the value of the translation.}
}
\description{
This generic method intends to extract the value of the symbol translation on the y of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = numeric}: method for \code{SpatialNetwork} objects.
}}

