% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spnet.title.list<-}
\alias{spnet.title.list<-}
\title{Set the list of all title parameters of a \code{SpatialNetwork} object}
\usage{
spnet.title.list(object) <- value
}
\arguments{
\item{object}{the \code{SpatialNetwork} object for which we want to set parameters.}

\item{value}{a list of parameters.}
}
\description{
This generic method intends to set or replace title parameters of a \code{SpatialNetwork} object.
}

