% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\docType{methods}
\name{spnet.title.list}
\alias{spnet.title.list}
\alias{spnet.title.list,SpatialNetwork-method}
\title{Get the list of all title parameters of a \code{SpatialNetwork} object}
\usage{
spnet.title.list(object)

\S4method{spnet.title.list}{SpatialNetwork}(object)
}
\arguments{
\item{object}{the \code{SpatialNetwork} object for which we want to get parameters.}
}
\description{
This generic method intends to extract title parameters of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.
}}

