% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{spnet.title.main<-}
\alias{spnet.title.main<-}
\title{Set the main title  of a \code{SpatialNetwork} object}
\usage{
spnet.title.main(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the new title.}
}
\description{
This generic method intends to set or replace the main title of a \code{SpatialNetwork} object.
}

