% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{spnet.title.main}
\alias{spnet.title.main}
\alias{spnet.title.main,SpatialNetwork-method}
\alias{spnet.title.main<-,SpatialNetwork,character-method}
\title{Get the main title of a \code{SpatialNetwork} object}
\usage{
spnet.title.main(object)

\S4method{spnet.title.main}{SpatialNetwork}(object)

\S4method{spnet.title.main}{SpatialNetwork,character}(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the new title.}
}
\description{
This generic method intends to extract the main title of a \code{SpatialNetwork} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SpatialNetwork}: method for \code{SpatialNetwork} objects.

\item \code{object = SpatialNetwork,value = character}: method for \code{SpatialNetwork} objects.
}}

