% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spnet.title.sub<-}
\alias{spnet.title.sub<-}
\title{Set the sub title  of a \code{SpatialNetwork} object}
\usage{
spnet.title.sub(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the new title.}
}
\description{
This generic method intends to set or replace the sub title of a \code{SpatialNetwork} object.
}

