\name{spnet.version}
\alias{spnet.version}
\title{
Get your current \code{spnet} package version
}
\description{
The \code{spnet.version} function return the version of the \code{spnet} package you are currently running.
}
\usage{
spnet.version()
}
\value{
A \code{character}.
}
\author{
Emmanuel Rousseaux
}
\examples{
spnet.version()
}
\keyword{ package }
\keyword{ version }
