% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{spnet.barplot.bound.lower<-}
\alias{spnet.barplot.bound.lower<-}
\title{Set the barplot lower bound position  of a \code{SpatialNetwork} object}
\usage{
spnet.barplot.bound.lower(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a numeric vector of coordinates, c(x,y), specifying a shift from the center of each country.}
}
\description{
This generic method intends to set or replace the barplot lower bound position of a \code{SpatialNetwork} object.
}

