% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spnet.barplot.variable<-}
\alias{spnet.barplot.variable<-}
\title{Set the barplot variable  of a \code{SpatialNetwork} object}
\usage{
spnet.barplot.variable(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the name of the variable to use for plotting barplots.}
}
\description{
This generic method intends to add or replace the barplot variable of a \code{SpatialNetwork} object.
}

