% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spnet.barplot.width<-}
\alias{spnet.barplot.width<-}
\title{Set the barplot width  of a \code{SpatialNetwork} object}
\usage{
spnet.barplot.width(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a numeric.}
}
\description{
This generic method intends to set or replace the barplot width of a \code{SpatialNetwork} object.
}

