% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{spnet.label.cex<-}
\alias{spnet.label.cex<-}
\title{Set the label cex  of a \code{SpatialNetwork} object}
\usage{
spnet.label.cex(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{numeric; the cex parameter.}
}
\description{
This generic method intends to set or replace the label cex of a \code{SpatialNetwork} object.
}

