% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialNetwork.r
\name{spnet.legend.line.width<-}
\alias{spnet.legend.line.width<-}
\title{Set the legend line width parameter of a \code{SpatialNetwork} object}
\usage{
spnet.legend.line.width(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a logical.}
}
\description{
This generic method intends to set or replace the legend line width parameter of a \code{SpatialNetwork} object.
}

