% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spnet.legend.ncol<-}
\alias{spnet.legend.ncol<-}
\title{Set the legend number of columns of a \code{SpatialNetwork} object}
\usage{
spnet.legend.ncol(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{a numeric.}
}
\description{
This generic method intends to add or replace the legend number of columns of a \code{SpatialNetwork} object.
}

