% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spnet.network.list<-}
\alias{spnet.network.list<-}
\title{Set the list of all parameters of a given network of a \code{SpatialNetwork} object}
\usage{
spnet.network.list(object, network.name) <- value
}
\arguments{
\item{object}{the \code{SpatialNetwork} object for which we want to set parameters.}

\item{network.name}{character; the name of the network.}

\item{value}{a list of parameters.}
}
\description{
This generic method intends to set or replace all parameters of a given network of a \code{SpatialNetwork} object.
}

