\docType{package}
\name{spnet}
\alias{spnet}
\alias{spnet-package}
\title{Plotting social networks on maps}
\description{
  The spnet package offers methods for dealing with spacial
  social networks. It allows to plot networks for which
  actors have a specific location on a map (participants in
  a political debate, cities, etc.). SpatialPolygons
  objects from the sp package are supported.
}
\details{
  \tabular{ll}{ Package: \tab spnet\cr Date: \tab
  2014-01-29\cr License: \tab GPL-3\cr }
}
\author{
  Emmanuel Rousseaux, Marion Deville \cr Maintainer:
  Emmanuel Rousseaux <emmanuel.rousseaux
}
\references{
  Deville M. and Rousseaux E. (2013). TITLE. JOURNAL.
  VOLUME(NUMBER), PAGEFROM-PAGETO.
}
\keyword{map}
\keyword{networks}
\keyword{package}
\keyword{spatial}
\keyword{spnet}

