% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spnet.r
\docType{package}
\name{spnet}
\alias{spnet}
\alias{spnet-package}
\title{Plotting social networks on maps}
\description{
The spnet package offers methods for dealing with spacial social networks. It allows to plot networks for which actors have a specific location on a map (participants in a political debate, cities, etc.). SpatialPolygons objects from the sp package are supported.
}
\references{
Rousseaux E., Deville M. and Ritschard G. (2014), "The SPNET package: Plotting social networks on maps with R", In 3èmes Rencontres R, Montpellier, France, June 25-27th.
}
\keyword{map}
\keyword{networks}
\keyword{package}
\keyword{spatial}
\keyword{spnet}

