% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spnet.symbol.legend<-}
\alias{spnet.symbol.legend<-}
\title{Set the symbol legend  of a \code{SpatialNetwork} object}
\usage{
spnet.symbol.legend(object) <- value
}
\arguments{
\item{object}{a \code{SpatialNetwork} object.}

\item{value}{the new legend.}
}
\description{
This generic method intends to add or replace the symbol legend of a \code{SpatialNetwork} object.
}

