\name{spnet.version}
\alias{spnet.version}
\title{Get your current \code{spnet} package version}
\usage{
  spnet.version()
}
\description{
  The \code{spnet.version} function return the version of
  the \code{spnet} package you are currently running.
}
\examples{
spnet.version()
}
\author{
  Emmanuel Rousseaux
}

